+++ axelor-account/src/main/java/com/axelor/apps/account/web/BudgetController.java
@@ -23,6 +23,8 @@ import com.axelor.apps.account.db.Budget;
 import com.axelor.apps.account.db.BudgetLine;
 import com.axelor.apps.account.db.repo.BudgetRepository;
 import com.axelor.apps.account.service.BudgetService;
+import com.axelor.exception.AxelorException;
+import com.axelor.exception.service.TraceBackService;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
@@ -46,7 +48,11 @@ public class BudgetController {
 	}
 	
 	public void generatePeriods(ActionRequest request, ActionResponse response) {
-		Budget budget = request.getContext().asType(Budget.class);
-		response.setValue("budgetLineList", budgetService.generatePeriods(budget));
+    try {
+      Budget budget = request.getContext().asType(Budget.class);
+      response.setValue("budgetLineList", budgetService.generatePeriods(budget));
+    } catch (AxelorException e) {
+      TraceBackService.trace(response, e);
+    }
 	}
 }