+++ axelor-base/src/main/java/com/axelor/apps/base/exceptions/IExceptionMessage.java
@@ -70,6 +70,7 @@ public interface IExceptionMessage {
 	
 	static final public String PERIOD_1 = /*$$(*/ "No period found or it has been closed for the company %s" /*)*/ ;
 	static final public String PERIOD_2 = /*$$(*/ "Period closed" /*)*/ ;
+  static final public String PERIOD_3 = /*$$(*/ "Too much iterations." /*)*/ ;
 	static final String PAY_PERIOD_CLOSED = /*$$(*/ "Warning : the pay period %s is closed." /*)*/;
 
 	/**