+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Accounting export, fix problem on export FEC
 - Fix readonly on date field in Fiscal year view form when creating a new record after having created a previous.
 - Fiscal Position interface reworked and moved from Account to Base module
+- Add a counter to limit iterations of while loop in YearService and BudgetService
+- Create an interface for PeriodService
 
 ## [4.2.3] - 2018-02-28
 4.1.3 with axelor-process-studio modules